added plugin system and separation of concerns
Created a plugin system for services similar to the one used by rdflib.
This allows people who just want to use the core library to do so
without pulling in a large number of dependencies. Deferred loading
means that they will only get an error if they try to pass in a service
that they did not install the dependencies for. (This isnt quite true.)

Split __init__.py into 3 major parts and 2 helpers (exceptions, utils).
The parts are terms, query, and services.

Terms provies the primary API for interacting with ontology identifiers
and interacting with metadata.

Query provides a layer of abstraction between terms and services so that
terms can transparently make use of multiple services. Query can also be
used directly to interface with multiple services, but the QueryResult
object is not streamlined as OntTerm, and there is validation.

Services provides the parent class with placeholder methods that all
services need to implement. It also holds a dependency free BasicService
which can be used to work with and store a small set of terms locally.