interlex service set self as source
the mini rdflib graph that was used as an intermediate local store
was being set at the source and you would have had to look at the
owl:Ontology header to see where it came from. This way we don't
inject the subgraph in. At some point it may still be useful to
keep that local graph around as well, but since everything is
converted to the ont term representation anyway it should be ok.